Add files via upload
So u need mediapipe and opencv...I have just made the code of writing on a black canvas...i am currently making a dataset to train a model and that picture drawn on canvas will be passed in that model and then it will show which letter I have drawn. This is not complete yet. My job was to write in the air and the letter will be displayed in a box automatically...but lets see how this turns out...